Make zbot3 withstand service crashes.
In the event that a service throws an error/exception during handling of an
event, an error report will be dumped to standard out. The error report will
contain the name of the offending service, the IRC event that triggered the
error, the current time, and the error that was thrown.