(closes #18) Implement stalker handler
The current implementation has the bot querying every joined channel
when anybody on any channel the bot is in QUITs, PARTs, NICKs, JOINs,
or MODEs. This is slightly inefficient, but whatever.

This fixes the bug where the !oprah command did not op people who were
in the channel before the bot joined.