Add Zbot.Extras.Color
The Color module provides an API for pretty printing colorized
messages. To use the developer pases a list of `ColorText` to the
`colorize` method which spits out an ANSI encoded colorized `Text`.

Example usage:

    colorize [
            bold $ bg Yellow $ fg White $ "[WARNING]"
        ,   " Some warning log message."
        ]

The `ColorText` type is of the `IsString` class so plain strings and
colorized strings can be mixed and matched within a list passed to
colorize.

Issue #38