Add Prometheus metrics
The prometheus-client library is used for collecting metrics and warp is
used to host the metrics endpoint.

The initial set of metrics include IRC I/O, errors, and the execution
time of the various services. For GHC metrics to be measured the command
line flags +RTS -T must be passed at runtime.

Closes #161