Only scrape text links (Take 2)
This change updates the web scraping utilities to silently truncate any
non-text mime type response. This should hopefully prevent resource
exhaustion when zbot encounters large images or media files...

In order to do this, scalpel was updated to 0.6, which meant bumping the
resolver to the latest LTS. This caused a GHC version bump which needed
some massaging to take care of.

Issue #174