Extend replies to support shouts and whispers
Many components use the onMessage family of utility functions to reply
to either a nick or a channel. There are some use cases (e.g. #120)
where the service may wish to respond via a whisper instead of a shout.

To support this a new ReplyMode type is defined and added as the first
parameter of the Reply function type. All uses of Reply have been
updated to use the Direct mode which maintains the current behavior.