Replace ReplyMode with MessageContext
MessageContext is a data type that can be extended in the future without
altering the type of all of the services that use the onMessage family
of functions.

This will allows ups in follow up changes to replace the usage of
onMessageWithChannel (#93) with MessageContext. With the added benefit
that all users of onMessage/onCommand now can get access to this
information for free.